A05/requirements.txt
rdkit
openbabel-wheel

